+++ src/protocols/packet_manager.cc
@@ -230,7 +230,7 @@ void PacketManager::decode(
             }
             else
             {
-                p->ip_proto_next = codec_data.next_prot_id;
+                p->ip_proto_next = (uint8_t)codec_data.next_prot_id;
             }
 
             fpEvalIpProtoOnlyRules(p, p->ip_proto_next);
@@ -391,7 +391,7 @@ bool PacketManager::encode(const Packet* p,
         flags |= ENC_FLAG_INLINE;
 
     ip::IpApi tmp_api;
-    EncState enc(tmp_api, flags, (uint16_t) next_prot, ttl, p->dsize);
+    EncState enc(tmp_api, flags, next_prot, ttl, p->dsize);
 
 
     const Layer* const lyrs = p->layers;
@@ -641,7 +641,7 @@ int PacketManager::encode_format_with_daq_info (
     int i;
     Layer* lyr;
     int len;
-    int num_layers = p->num_layers;
+    uint8_t num_layers = p->num_layers;
     DAQ_PktHdr_t* pkth = (DAQ_PktHdr_t*)c->pkth;
 
     if ( num_layers <= 0 )
